# static/swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "API de Geração de QR Code",
      "description": "API para gerar QR Codes e gerenciar usuários",
      "termsOfService": "",
      "contact": {
        "name": "Suporte",
        "url": "",
        "email": "suporte@example.com"
      }
    },
    "host": "127.0.0.1:5000",
    "basePath": "/",
    "schemes": [
      "http"
    ],
    "paths": {
      "/register": {
        "post": {
          "summary": "Registrar um novo usuário",
          "description": "Crie um novo usuário e obtenha uma APIKey",
          "consumes": ["application/json"],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Dados do usuário",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": ["username", "password"]
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Usuário registrado com sucesso",
              "schema": {
                "$ref": "#/definitions/User"
              }
            },
            "400": {
              "description": "Nome de usuário já existe"
            }
          }
        }
      },
      "/generate_qrcode": {
        "post": {
          "summary": "Gerar QR Code",
          "description": "Forneça um conteúdo para gerar um QR Code",
          "consumes": ["application/json"],
          "parameters": [
            {
              "in": "header",
              "name": "x-api-key",
              "type": "string",
              "required": true,
              "description": "API key para autenticação"
            },
            {
              "in": "body",
              "name": "body",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string"
                  }
                },
                "required": ["content"]
              },
              "description": "Conteúdo para ser transformado em QR Code"
            }
          ],
          "responses": {
            "200": {
              "description": "QR Code gerado com sucesso"
            },
            "400": {
              "description": "Conteúdo para QR Code ausente"
            },
            "401": {
              "description": "API key ausente ou inválida"
            }
          }
        }
      }
    },
    "definitions": {
      "User": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "api_key": {
            "type": "string"
          }
        }
      }
    }
  }
  